// project-management-frontend/src/app/services/models/page-response-project-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ProjectResponse } from '../models/project-response';
export interface PageResponseProjectResponse {
  content?: Array<ProjectResponse>;
  first?: boolean;
  last?: boolean;
  number?: number;
  size?: number;
  totalElements?: number;
  totalPages?: number;
}