// project-management-frontend/src/app/services/models/project-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ProjectMemberResponse } from '../models/project-member-response';
import { Task } from '../models/task';
export interface ProjectResponse {
  description?: string;
  id?: number;
  members?: Array<ProjectMemberResponse>;
  name?: string;
  owner?: string;
  startDate?: string;
  tasks?: Array<Task>;
}