// project-management-frontend/src/app/services/models/task-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface TaskResponse {
  assigned?: boolean;
  assignedTo?: string;
  description?: string;
  dueDate?: string;
  id?: number;
  name?: string;
  priority?: 'LOW' | 'MEDIUM' | 'HIGH';
  projectName?: string;
  status?: 'TODO' | 'IN_PROGRESS' | 'DONE';
}